refactor: move some code blocks to separate files
