+++ src/features/login-form/login-form.tsx
@@ -1,98 +1,16 @@
-import { useState } from 'react';
-import { Control, Controller, useForm } from 'react-hook-form';
+import { FC, useState } from 'react';
+import { useForm } from 'react-hook-form';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import Visibility from '@mui/icons-material/Visibility';
-import VisibilityOff from '@mui/icons-material/VisibilityOff';
-import { Alert, Button, Collapse, IconButton, InputAdornment, TextField } from '@mui/material';
+import { Alert, Button, Collapse } from '@mui/material';
 import { useMutation } from '@tanstack/react-query';
 
-import { getToken } from '@/lib/axios/requests/get-token';
-import { signInCustomer } from '@/lib/axios/requests/sign-in-customer';
+import { ControlledTextField } from '../../components/controlled-text-field';
+import { loginUser } from '../../lib/axios/requests/login-user';
+import { LoginInfo, loginSchema } from './login-form.schema.ts';
+import { PasswordInput } from './password-input.tsx';
 
-import { LoginInfo, loginSchema } from './login-form.schema';
-
-const ControllerComponent = ({
-  control,
-  label,
-  name,
-}: {
-  control: Control<LoginInfo>;
-  label: string;
-  name: keyof LoginInfo;
-}): JSX.Element => {
-  return (
-    <Controller
-      control={control}
-      name={name}
-      render={({ field, fieldState }) => {
-        return (
-          <TextField
-            error={fieldState.invalid}
-            helperText={fieldState.error?.message ?? ' '}
-            label={label}
-            size="small"
-            {...field}
-          />
-        );
-      }}
-    />
-  );
-};
-
-async function loginUser({
-  password,
-  username,
-}: {
-  password: string;
-  username: string;
-}): Promise<void> {
-  const tokenInfo = await getToken({ password, username });
-  const result = await signInCustomer({ email: username, password }, tokenInfo.access_token);
-  console.log(result);
-}
-
-const PasswordInput = (control: Control<LoginInfo>): JSX.Element => {
-  const [showPassword, setShowPassword] = useState(false);
-
-  const handleClickShowPassword = (): void => setShowPassword((show) => !show);
-
-  return (
-    <Controller
-      control={control}
-      name="password"
-      render={({ field, fieldState }) => {
-        return (
-          <TextField
-            InputProps={{
-              endAdornment: (
-                <InputAdornment position="end">
-                  <IconButton
-                    aria-label="toggle password visibility"
-                    edge="end"
-                    onClick={handleClickShowPassword}
-                    onMouseDown={(event) => event.preventDefault()}
-                    onMouseUp={(event) => event.preventDefault()}
-                  >
-                    {showPassword ? <VisibilityOff /> : <Visibility />}
-                  </IconButton>
-                </InputAdornment>
-              ),
-            }}
-            {...field}
-            error={fieldState.invalid}
-            helperText={fieldState.error?.message ?? ' '}
-            label="Password"
-            size="small"
-            type={showPassword ? 'text' : 'password'}
-          />
-        );
-      }}
-    />
-  );
-};
-
-export const LoginForm = (): JSX.Element => {
+export const LoginForm: FC = () => {
   const { control, handleSubmit } = useForm<LoginInfo>({
     defaultValues: { email: '', password: '' },
     mode: 'onChange',
@@ -118,7 +36,7 @@ export const LoginForm = (): JSX.Element => {
         })(e)
       }
     >
-      <ControllerComponent {...{ control, label: 'Email', name: 'email' }} />
+      <ControlledTextField control={control} label="Email" name="email" />
       <PasswordInput {...control} />
       {
         <Collapse in={!!errorMessage}>
